requirements.txt
flask
gunicorn[eventlet]
django
pyyaml